% Generated by roxygen2 (4.0.1): do not edit by hand
\name{addInteractions}
\alias{addInteractions}
\title{addInteractions: Add intercepts, interaction, and polynomial terms to dataframe}
\usage{
addInteractions(data)
}
\arguments{
\item{data}{Dataframe of covariates to model or predict daily stream temperature}
}
\value{
Original dataframe plus the additional terms
}
\description{
\code{addInteractions} Add intercepts, interaction, and polynomial terms to dataframe
}
\details{
By making this a function it can easily get added to multiple places in the data prep for validation, fitting, or predicting. Then there is only one place where the interactions have to be specified
}
\examples{
\dontrun{
tempDataSyncS <- addInteractions(tempDataSyncS)
}
}

