% Generated by roxygen2 (4.0.2): do not edit by hand
\name{addStreamMuResid}
\alias{addStreamMuResid}
\title{addStreamMuResid}
\usage{
addStreamMuResid(M.wb, tempDataSyncS)
}
\arguments{
\item{M,tempDataSyncS}{}
}
\description{
\code{addStreamMuResid} get means for stream.mu and resid from the jags output. append means to tempDataSyncS
}
\details{
var: blah, blah, blah
value: something, something
}

