% Generated by roxygen2 (4.0.1): do not edit by hand
\name{avgCoefs}
\alias{avgCoefs}
\title{avgCoefs: Calculate the mean coefficient estimates from the MCMC iterations}
\usage{
avgCoefs(ggs.obj, family = NULL)
}
\arguments{
\item{ggs.obj}{An object generated with the ggmcmc::ggs() function that organizes the jags output mcmc.list}

\item{family}{Optional character string specifying the subset of the parameters to average (e.g. "B.0" for parameters B.0[1], B.0[2], B.0[3], etc.)}
}
\description{
\code{avgCoefs} returns a dplyr tbl_df with the mean, sd, and 95% credible intervals for each parameter
}
\details{
blah, blah, blah
}

