% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Derive_Metrics.R
\name{calc7DADM}
\alias{calc7DADM}
\title{calc7DADM}
\usage{
calc7DADM(grouped.df, derived.df, threshold)
}
\arguments{
\item{grouped.df}{Dataframe grouped by featureid then year}

\item{derived.df}{Dataframe of derived metrics}

\item{threshold}{Temperature threshold above which causes physiological problems for species of interest}
}
\value{
Returns Dataframe of derived metrics (e.g. max temperature predicted over daymet record) for each featureid across all years
}
\description{
\code{calc7DADM} Calculates the 7-day average of the daily (max) temperatures in any 7-contiguous day period from June - September
}
\details{
This is part of MassDEP CALM criteria for evaluating Cold Water Streams for 305b reporting. MassDEP uses 20C as their threshold ("endpoint").
}
\examples{
\dontrun{

}
}

