% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Derive_Metrics.R
\name{calcConsecExceed}
\alias{calcConsecExceed}
\title{calcConsecExceed}
\usage{
calcConsecExceed(y, threshold = 22)
}
\arguments{
\item{threshold}{Temperature threshold above which causes physiological problems for species of interest}

\item{grouped.df}{Dataframe grouped by featureid then year}

\item{derived.df}{Dataframe of derived metrics}

\item{summer}{Logical calculate metric just for June-August. Default FALSE}
}
\value{
Returns Dataframe of derived metrics (e.g. max temperature predicted over daymet record) for each featureid across all years
}
\description{
\code{calcConsecExceed} Calculate the mean and max (and quantiles?) of consecutive days per year where temperatures exceed some temperature threshold
}
\details{
Maine is interested in adding another set of derived metrics that represented runs of consecutive days where the stream temperature exceeded some threshold. Maybe something like mean and max number of consecutive days per year where temp > 22 deg C, since that's closely tied to fish stress.
}
\examples{

\dontrun{

}
}

