% Generated by roxygen2 (4.0.1): do not edit by hand
\name{calcYearsCold}
\alias{calcYearsCold}
\title{calcYearsCold}
\usage{
calcYearsCold(grouped.df, derived.df, temp.threshold, states)
}
\arguments{
\item{grouped.df}{Dataframe grouped by featureid then year}

\item{derived.df}{Dataframe of derived metrics}

\item{state}{Character string for specific State}
}
\value{
Returns Dataframe of derived metrics (e.g. max temperature predicted over daymet record) for each featureid across all years
}
\description{
\code{calcYearsCold} Calculate derived metrics from predicted stream temperatures
}
\details{
Different states have different classifications of cold, cool, and warm streams for regulatory purposes. This function takes the state abreviation and returns the number and frequency of years that any stream reach is expected to fall within each category.
}
\examples{
\dontrun{

}
}

