% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Derive_Metrics.R
\name{calcYearsMaxTemp}
\alias{calcYearsMaxTemp}
\title{calcYearsMaxTemp}
\usage{
calcYearsMaxTemp(grouped.df, derived.df, temp.threshold, summer = FALSE)
}
\arguments{
\item{grouped.df}{Dataframe grouped by featureid then year}

\item{derived.df}{Dataframe of derived metrics}

\item{temp.threshold}{Optional numeric temperature threshold value in degrees C}
}
\value{
Returns Dataframe of derived metrics (e.g. max temperature predicted over daymet record) for each featureid across all years
}
\description{
\code{calcYearsMaxTemp} Calculates the number of years and frequency of years that a stream ever (even for 1 day) exceeds a threshold temperature
}
\details{
blah, blah, blah, something, something
}
\examples{

\dontrun{

}
}

