% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conteStreamTemperature-package.r
\docType{package}
\name{conteStreamTemperature}
\alias{conteStreamTemperature}
\alias{conteStreamTemperature-package}
\title{conteStreamTemperature.}
\description{
conteStreamTemperature.
}

