% Generated by roxygen2 (4.0.2): do not edit by hand
\name{convert_neg}
\alias{convert_neg}
\title{convert_neg}
\usage{
convert_neg(data)
}
\arguments{
\item{data}{time series data pull from the postgres database (df_values table)}
}
\value{
temp Convert -1 - 0 C readings to 0
}
\description{
\code{convert_neg} Check for excessively cold events
}
\details{
Convert -1 - 0 C readings to 0
}
\examples{
\dontrun{

}
}

