% Generated by roxygen2 (4.0.1): do not edit by hand
\name{correctCovariateData}
\alias{correctCovariateData}
\title{correctCovariateData}
\usage{
correctCovariateData(covariateData, siteChanges, LocalStats, UpstreamStats,
  impoundmentLayers)
}
\arguments{
\item{covariateData}{The original covariate data file}

\item{siteChanges}{The "siteChanges" CSV file}

\item{LocalStats}{The master dataframes of both local and upstream covariate statistics}

\item{UpstreamStats}{from localstats?????}

\item{impoundmentLayers}{A list of layers that get NAs assigned for local catchment values. (This will hopefully change with an updated layer)}
}
\value{
the same covariate dataframe with corrected values and a column indicating whether or not values
}
\description{
\code{correctCovariateData} corrects the covariate data file after the site locations have been manually checked.
}
\details{
This function corrects the covariate data file after the site locations have been manually checked.
}

