% Generated by roxygen2 (4.0.1): do not edit by hand
\name{createEvalRows}
\alias{createEvalRows}
\title{createEvalRows: Create rows to loop through for autoregressive function}
\usage{
createEvalRows(data)
}
\arguments{
\item{data}{Dataframe for analysis with date and a deployID row (can generate with indexDeployments function)}
}
\description{
\code{createDeployRows} returns a list of two dataframes each with a rowNum column for looping
}
\details{
var: blah, blah, blah
value: something, something
}

