% Generated by roxygen2 (4.0.2): do not edit by hand
\name{derive_metrics_par}
\alias{derive_metrics_par}
\title{derive_metrics_par}
\usage{
derive_metrics_par(i, chunk.size = 50, catchmentid, springFallBPs,
  tempDataSync, df_covariates_upstream, featureid_huc8,
  featureid_site = featureid_site, featureid_lat_lon, coef.list, cov.list,
  var.names)
}
\arguments{
\item{i}{Iteration passed from foreach loop}

\item{chunk.size}{Number of catchments to process at 1 time, default = 50 based on testing speed with pulling from the database and memory issues}

\item{catchmentid}{}

\item{springFallBPs}{}

\item{tempDataSync}{}

\item{df_covariates_upstream}{}

\item{featureid_huc8}{}

\item{featureid_lat_lon}{}

\item{coef.list}{}

\item{cov.list}{}

\item{var.names}{}

\item{featureid_site=featureid_site}{}
}
\value{
Returns Dataframe of derived metrics (e.g. max temperature predicted over daymet record) for each featureid across all years
}
\description{
\code{derive_metrics_par} Wrapper to pull data from database, prepare data, and calculate derived metrics in parallel using foreach %dopar%
}
\details{
blah, blah, blah, something, something
}
\examples{
\dontrun{

}
}

