% Generated by roxygen2 (4.0.2): do not edit by hand
\name{deriveMetrics}
\alias{deriveMetrics}
\title{deriveMetrics}
\usage{
deriveMetrics(data, threshold = NULL)
}
\arguments{
\item{data}{Dataframe that includes at least site, year, temp (observed), and tempPredicted}
}
\value{
Returns Dataframe of derived metrics (e.g. max temperature predicted over daymet record) for each site across all years
}
\description{
\code{deriveMetrics} Calculate derived metrics from predicted stream temperatures
}
\details{
This function calculates the following metrics based on predicted stream temperature across years

* total observations (days with data) per site
* Mean maximum daily mean temperature by site (over years)
* Maximum max daily mean temperature
* Number of days with stream temp > 18, 20, 22 C and optional user-defined temperature
* Number of years with mean maximum over 18, 20, 22 C and optional user-defined temperature
* frequency of years with a mean max over 18, 20, 22 C and optional user-defined temperature
* Mean resistance to peak air temperature (difference between observed air and predicted stream temperatures during the hottest part of the year)
* Mean RMSE for each site
* Flag based on RMSE > 95%. These sites should probably be checked for unrecorded impoundments, restrictive culverts, or large groundwater influences before making management or policy decisions
}
\examples{
\dontrun{
df <- predictTemp(data = df, data.fit = tempDataSyncS, cov.list = cov.list, coef.list = coef.list)
derived.metrics <- deriveMetrics(data = df)

}
}

