% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Derive_Metrics.R
\name{deriveMetricsYear}
\alias{deriveMetricsYear}
\title{deriveMetricsYear}
\usage{
deriveMetricsYear(data)
}
\arguments{
\item{data}{Dataframe created with the prepData and predictTemp functions}
}
\value{
Returns Dataframe of derived metrics (e.g. max temperature predicted over daymet record) for each featureid across all years
}
\description{
\code{deriveMetricsYear} Wrapper to calculate derived metrics in parallel using foreach %dopar%
}
\details{
blah, blah, blah, something, something
}
\examples{

\dontrun{

}
}

