% Generated by roxygen2 (4.0.1): do not edit by hand
\name{findNearestDaymetPoint}
\alias{findNearestDaymetPoint}
\title{findNearestDaymetPoint}
\usage{
findNearestDaymetPoint(siteLat, siteLon, dayLat, dayLon, currentTile)
}
\arguments{
\item{siteLat}{A single latitude value of the site.}

\item{siteLon}{A single longitude value of the site.}

\item{dayLat}{An array of latitude points as read in from Daymet NetCDF file.}

\item{dayLon}{An array of longitude points as read in from Daymet NetCDF file.}

\item{currentTile}{The Daymet tile that is currently being indexed.}
}
\value{
Returns array position (row, col) of the nearest Daymet point to the site.
}
\description{
\code{findNearestDaymetPoint} This function takes the latitude and longitude of a site and calculates the nearest Daymet point based on the latitude/longitude arrays read in from Daymet.
}
\details{
This function takes the latitude and longitude of a site and calculates the nearest Daymet point based on the latitude/longitude arrays read in from Daymet. It returns the array position (row, col) of the nearest Daymet point to the site. This is used to index the climate records from the variable array as read in from Daymet.
}
\examples{
\dontrun{
findNearestDaymetPoint(siteLat = 44.0 , siteLon = -72.5, dayLat = lat, dayLon = lon, currentTile = 11934)
}
}

