% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QAQC.R
\name{flag_cold_days}
\alias{flag_cold_days}
\title{flag_cold_days}
\usage{
flag_cold_days(data)
}
\arguments{
\item{data}{time series data pull from the postgres database (df_values table)}
}
\value{
Flagging days < -1 C
}
\description{
\code{flag_cold_days} Flagging days < -1 C
}
\details{
Flagging days < -1 C
}
\examples{

\dontrun{

}
}

