% Generated by roxygen2 (4.0.2): do not edit by hand
\name{flag_extreme_obs}
\alias{flag_extreme_obs}
\title{flag_extreme_obs}
\usage{
flag_extreme_obs(data, qlo = 0.001, qhi = 0.999)
}
\arguments{
\item{data}{time series data pull from the postgres database (df_values table)}

\item{qlo}{Lower quantile to flag. Default = 0.001}

\item{qhi}{Upper quantile to flag. Default = 0.999}
}
\value{
logical Flag for extreme observations
}
\description{
\code{flag_extreme_obs} Flagging extreme observations
}
\details{
Flagging upper and lower 5th percentiles probably wont work for the database, as those are best reverified by whoever collected/uploaded the data.
}
\examples{
\dontrun{

}
}

