% Generated by roxygen2 (4.0.2): do not edit by hand
\name{flag_hot_days}
\alias{flag_hot_days}
\title{flag_hot_days}
\usage{
flag_hot_days(data, threshold = 35)
}
\arguments{
\item{data}{time series data pull from the postgres database (df_values table)}

\item{threshold}{Temperature threshold to flag. Default = 35}
}
\value{
logical Flag for hot days
}
\description{
\code{flag_hot_days} Flagging hot days
}
\details{
Flagging observations > 30 C might not work for the database, as it isn't all small trout streams.
}
\examples{
\dontrun{

}
}

