% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QAQC.R
\name{flag_hot_obs}
\alias{flag_hot_obs}
\title{flag_hot_obs}
\usage{
flag_hot_obs(data, threshold = 35)
}
\arguments{
\item{data}{time series data pull from the postgres database (df_values table)}

\item{threshold}{Temperature threshold to flag. Default = 35}
}
\value{
logical Flag for hot observations
}
\description{
\code{flag_hot_obs} Flagging hot observations
}
\details{
Flagging observations > 30 C might not work for the database, as it isn't all small trout streams.
}
\examples{

\dontrun{

}
}

