% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/QAQC.R
\name{flag_hourly_rise}
\alias{flag_hourly_rise}
\title{flag_hourly_rise}
\usage{
flag_hourly_rise(data, deg = 5)
}
\arguments{
\item{data}{time series data pull from the postgres database (df_values table)}

\item{deg}{Threshold change in degrees for flag}
}
\value{
Flag days with Rate of change in temperature per hour above deg
}
\description{
\code{flag_hourly_rise} Check for out of water (out of water vs. dry stream?)
}
\details{
Rate of change in temperature per hour
}
\examples{
\dontrun{

}
}

