% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QAQC.R
\name{flag_interval}
\alias{flag_interval}
\title{flag_interval}
\usage{
flag_interval(data)
}
\arguments{
\item{data}{time series data pull from the postgres database (df_values table)}
}
\value{
flag inconsistent time intervals within a day at a given location or series (logger deployment)
}
\description{
\code{flag_interval} Uniformity of records within a day
}
\details{
check to make sure datetime was done correctly when guessing 12-hr vs 24-hr formats
}
\examples{

\dontrun{

}
}

