% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Data_Prep_Functions.R
\name{\%!in\%}
\alias{\%!in\%}
\title{not in:}
\usage{
x \%!in\% table
}
\arguments{
\item{x}{}

\item{table}{}
}
\description{
\code{!in} antii %in%
}
\details{
var: blah, blah, blah
value: something, something
}

