% Generated by roxygen2 (4.0.1): do not edit by hand
\name{indexCovariateData}
\alias{indexCovariateData}
\title{indexCovariateData}
\usage{
indexCovariateData(record, masterCovariates, catchmentShapefile,
  projectionString, fields)

indexCovariateData(record, masterCovariates, catchmentShapefile,
  projectionString, fields)
}
\arguments{
\item{record}{The stream temperature record (unique site ID, latitude, and longitude columns )}

\item{masterCovariates}{A dataframe of the covariates for the catchments (FEATUREIDs source)}

\item{catchmentShapefile}{A master catchments shapefile}

\item{projectionString}{A CRS string of the spatial data projection}

\item{fields}{A string of variables to pull from the covariates list}
}
\value{
Returns a dataframe with the site name, lat/lon, FEATUREID, and the select covariate values
}
\description{
\code{indexCovariateData} Indexes values from the master list of covariates for observed stream temperature sites

\code{indexCovariateData} returns something of great importance
}
\details{
This function indexes values from the master list of covariates for observed stream temperature sites.

var: blah, blah, blah
value: something, something
This function indexes values from the master list of covariates for observed stream temperature sites.
It takes the following:
   1) The stream temperature record (unique site ID, latitude, longitude, columns)
   2) A dataframe of the covariates for the catchments (FEATUREIDs source)
   3) A master catchments shapefile
   4) A CRS string of the spatial data projection
   5) A string of variables to pull from the covariates list

It returns a dataframe with the site name, lat/lon, FEATUREID, and the select covariate values.
}

