% Generated by roxygen2 (4.0.1): do not edit by hand
\name{indexDaymetTileByLatLon}
\alias{indexDaymetTileByLatLon}
\title{indexDaymetTileByLatLon}
\usage{
indexDaymetTileByLatLon(SiteLat, SiteLon)
}
\arguments{
\item{SiteLat}{Site latitude}

\item{SiteLon}{Site longitude}
}
\value{
Returns the Daymet tile that holds data for the coordinates given.
}
\description{
\code{indexDaymetTileByLatLon} This function indexes Daymet tiles by the latitude and longitude of a site. sites
}
\details{
This function indexes Daymet tiles by the latitude and longitude of a site.
}

