% Generated by roxygen2 (4.0.1): do not edit by hand
\name{indexDeployments}
\alias{indexDeployments}
\title{indexDeployments: Create an index for each unique temperature logger deployment}
\usage{
indexDeployments(data, regional = FALSE)
}
\arguments{
\item{data}{Dataframe of temperature data including date and unique site ID}

\item{regional}{Optional logical if true then also sort the dataframe by HUC8}
}
\description{
\code{indexDeployments} returns the input dataframe with a deployment ID added as a column
}
\details{
var: blah, blah, blah
value: something, something
}

