% Generated by roxygen2 (4.0.2): do not edit by hand
\name{indexLocalDaymetVariables}
\alias{indexLocalDaymetVariables}
\title{indexLocalDaymetVariables}
\usage{
indexLocalDaymetVariables(record, variables, daymetDirectory)
}
\arguments{
\item{record}{dataframe indicating the period of record ( limited to 1980 - 2013) over which to pull climate data. Dataframe columns must be in the following order: 1)Site ID 2)Year 3)Day of Year 4)Longitude 5)Latitude}

\item{variables}{character string listing the climate variables (as named by Daymet) to be indexed}

\item{daymetDirectory}{character string indicating the folder where Daymet files reside}
}
\value{
Returns the record with added columns for specified Daymet variable records.
}
\description{
\code{indexLocalDaymetVariablesForObservedSites} takes a period of record for a location and returns the corresponding time series of climate data from Daymet. This is done by indexing the nearest Daymet point to the site location.
}
\details{
This function takes a period of record for a location and returns the corresponding time series of climate data from Daymet. This is done by indexing the nearest Daymet point to the site location.
Input names for "record" can vary as long as order is correct. Original names will be included in output.
}
\examples{
\dontrun{
indexLocalDaymetVariables(record = masterData , variables = c('prcp', 'tmin', 'tmax'), daymetDirectory = '//IGSAGBEBWS-MJO7/projects/dataIn/environmental/climate/daymet/unzipped/Daily')
}
}

