% Generated by roxygen2 (4.0.1): do not edit by hand
\name{indexLocalDaymetVariablesForObservedSites}
\alias{indexLocalDaymetVariablesForObservedSites}
\title{indexLocalDaymetVariablesForObservedSites}
\usage{
indexLocalDaymetVariablesForObservedSites(record, variables)
}
\arguments{
\item{record}{dataframe indicating the period of record ( limited to 1980 - 2013) over which to pull climate data.}

\item{variables}{A string listing the climate variables (as named by Daymet) to be indexed}

\item{daymetDirectory}{string indicating the folder where Daymet files reside}
}
\value{
Returns the record with added columns for Daymet variable records.
}
\description{
\code{indexLocalDaymetVariablesForObservedSites} takes a period of record for a location and returns the corresponding time series of climate data from Daymet. This is done by indexing the nearest Daymet point to the site location.
}
\details{
This function takes a period of record for a location and returns the corresponding time series of climate data from Daymet. This is done by indexing the nearest Daymet point to the site location.

Arguments:
   1) record          A dataframe indicating the period of record ( limited to 1980 - 2013) over which to
                        pull climate data. Minimum dataframe column requirements and format:
                        ---------------------------------------------------------------------------------
                        'data.frame':  3228940 obs. of  5 variables:
                         $ site     : chr  "MADEP_W1466_M1" "MADEP_W1466_M1" "MADEP_W1466_M1" ...
                         $ year     : num  1980 1980 1980 1980 1980 1980 1980 1980 1980 ...
                         $ dOY      : num  1 2 3 4 5 6 7 8 9 ...
                         $ Latitude : num  42.5 42.5 42.5 42.5 ...
                         $ Longitude: num  -72.9 -72.9 -72.9 -72.9 ...
                        ---------------------------------------------------------------------------------
   2) variables       A string listing the climate variables (as named by Daymet) to be indexed
   3) daymetDirectory A string indicating the folder where Daymet files reside
}
\examples{
\dontrun{
indexLocalDaymetVariablesForObservedSites(record = masterData , variables = c('prcp', 'tmin', 'tmax'), daymetDirectory = '//IGSAGBEBWS-MJO7/projects/dataIn/environmental/climate/daymet/unzipped/Daily')
}
}

