% Generated by roxygen2 (4.0.2): do not edit by hand
\name{indexUpstreamDaymetVariablesByCatchment}
\alias{indexUpstreamDaymetVariablesByCatchment}
\title{indexUpstreamDaymetVariablesByCatchment}
\usage{
indexUpstreamDaymetVariablesByCatchment(record, variables, catchmentShapefile,
  catchmentID, delineatedCatchmentsList, daymetDirectory)
}
\arguments{
\item{record}{dataframe indicating the period of record (limited to 1980 - 2013) over which to pull climate data. Columns must be in the following order: 1)Catchment ID 2)Year 3)Day of Year 4)Site ID (if applicable)}

\item{variables}{A string of variables to pull from Daymet. Options: "prcp", "tmin", "tmax", "srad", "vp", "swe", & "dayl"}

\item{catchmentShapefile}{SpatialPolygonsDataframe of the catchment shapefile covering the area over which the sites will be delineated}

\item{catchmentID}{a character vector of column name describing the catchment identifier in the shapefile (e.g. "FEATUREID" for NHDplusV2)}

\item{delineatedCatchmentsList}{A list of catchment delineations for the region}

\item{daymetDirectory}{Directory}
}
\value{
Returns the original dataframe with new columns for the Daymet variables.
}
\description{
\code{indexUpstreamDaymetVariablesByCatchment} calculates the spatial average of the Daymet variables within a watershed.
}
\details{
This function calculates the spatial average of the Daymet variables within a watershed.
}

