% Generated by roxygen2 (4.0.1): do not edit by hand
\name{indexUpstreamDaymetVariablesForObservedSites}
\alias{indexUpstreamDaymetVariablesForObservedSites}
\title{indexUpstreamDaymetVariablesForObservedSites}
\usage{
indexUpstreamDaymetVariablesForObservedSites(record, variables, tiles,
  catchmentShapefile, covariateData, delineatedCatchmentsList)
}
\arguments{
\item{record}{The stream temperature record (Site names, latitude, longitude, year, and dOY columns)}

\item{variables}{A string of variables to pull from Daymet}

\item{tiles}{A list of daymet tiles covered by the watersheds}

\item{catchmentShapefile}{A master catchments shapefile}

\item{covariateData}{A dataframe of the covariates for the catchments (FEATUREIDs source)}

\item{delineatedCatchmentsList}{A list of catchment delineations for the region}

\item{daymetDirectory}{Directory}
}
\value{
Returns the original dataframe with new columns for the Daymet variables.
}
\description{
\code{indexUpstreamDaymetVariablesForObservedSites} calculates the spatial average of the Daymet variables within a watershed.
}
\details{
This function calculates the spatial average of the Daymet variables within a watershed.
}

