% Generated by roxygen2 (4.0.1): do not edit by hand
\name{loggAndScaleDaymet}
\alias{loggAndScaleDaymet}
\title{indexUpstreamDaymetVariablesForObservedSites}
\usage{
loggAndScaleDaymet(dataFrame, logVars, scaleVars)
}
\arguments{
\item{dataFrame}{dataframe}

\item{logVars}{somthing}

\item{scaleVars}{something}
}
\value{
Returns something
}
\description{
\code{indexUpstreamDaymetVariablesForObservedSites} This function takes a dataframe of time series data and lists of variables to either scale or log-scale.
}
\details{
This function takes a dataframe of time series data and lists of variables to either scale or log-scale.
}

