% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QAQC.R
\name{MAD.roller}
\alias{MAD.roller}
\title{MAD.roller}
\usage{
MAD.roller(vals, window)
}
\arguments{
\item{vals}{Numeric temperature values}

\item{window}{Integer number of points to include in the window calculating the median for comparision}
}
\value{
Median Absolute Deviation for each value compared to the windowed median
}
\description{
\code{MAD.roller} Median Absolute Deviations
}
\details{
Adapted from the sensorQC package
blah, blah, blah, something, something
}
\examples{

\dontrun{

}
}

