% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mae}
\alias{mae}
\title{mae: mean absolute error}
\usage{
mae(error)
}
\arguments{
\item{error}{Vector of residual error from a model}
}
\description{
\code{mae} returns mean absolute error
}
\details{
var: blah, blah, blah
value: something, something
}

