% Generated by roxygen2 (4.0.1): do not edit by hand
\name{modelRegionalTemp}
\alias{modelRegionalTemp}
\title{modelRegionalTemp}
\usage{
modelRegionalTemp(data = tempDataSyncS, params = c("sigma", "B.0", "B.site",
  "rho.B.site", "mu.site", "sigma.b.site", "B.year", "rho.B.year", "mu.year",
  "sigma.b.year"), n.burn = 5000, n.it = 3000, n.thin = 3, n.chains = 3,
  coda = FALSE, runParallel = TRUE)
}
\arguments{
\item{data}{dataframe created in the 3-statModelPrep.Rmd script}

\item{params}{Character string of parameters to monitor (return) from the model}

\item{n.burn}{Integer number of iterations in the burn-in (adapation) phase of the MCMC}

\item{n.it}{Integer number of iterations per chain to run after the burn-in}

\item{n.thin}{Integer: save every nth iteration}

\item{n.chains}{Integer number of chains. One run per cluster so should use <= # cores on computer}

\item{coda}{Logical if TRUE return coda mcmc.list, if FALSE (default) return jags.samples object}
}
\value{
Returns the iterations from the Gibbs sampler for each variable in params as either an mcmc.list or jags.samples object
}
\description{
\code{modelRegionalTemp} Linear mixed model in JAGS to model daily stream temperature
}
\details{
This function takes daily observed stream temperatures, air temperature, snow-water-equivalent (swe), day of the year, and landscape covariates for a linear mixed effects model with site and year random effects.
}
\examples{
\dontrun{
mcmc.out <- modelRegionalTemp(tempDataSyncS, n.burn = 1000, n.it = 1000, n.thin = 3, n.chains = 3)
}
}

