% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/jagsWrappers.R
\name{modelRegionalTempHUC}
\alias{modelRegionalTempHUC}
\title{modelRegionalTempHUC}
\usage{
modelRegionalTempHUC(data = tempDataSyncS, data.fixed, data.random.sites,
  data.random.years, param.list, n.burn = 5000, n.it = 3000, n.thin = 3,
  nc = 3, coda = FALSE, runParallel = TRUE)
}
\arguments{
\item{data}{dataframe created in the 3-statModelPrep.Rmd script}

\item{data.fixed}{Dataframe of fixed effects parameter names from columns in data. These are parameters without random slopes.}

\item{data.random.sites}{Dataframe of variables to have random slopes by site}

\item{data.random.years}{Dataframe of variables to have random slopes by year}

\item{n.burn}{Integer number of iterations in the burn-in (adapation) phase of the MCMC}

\item{n.it}{Integer number of iterations per chain to run after the burn-in}

\item{n.thin}{Integer: save every nth iteration}

\item{coda}{Logical if TRUE return coda mcmc.list, if FALSE (default) return jags.samples object}

\item{params}{Character string of parameters to monitor (return) from the model}

\item{n.chains}{Integer number of chains. One run per cluster so should use <= # cores on computer}
}
\value{
Returns the iterations from the Gibbs sampler for each variable in params as either an mcmc.list or jags.samples object
}
\description{
\code{modelRegionalTempHUC} Linear mixed model in JAGS to model daily stream temperature
}
\details{
This function takes daily observed stream temperatures, air temperature, day of the year, and landscape covariates for a linear mixed effects model with site within HUC8 and year random effects.
}
\examples{
\dontrun{
M.huc <- modelRegionalTempHUC(data, data.fixed, data.random.sites, data.random.years, n.burn = 1000, n.it = 1000, n.thin = 1, nc = 3, coda = coda.tf, param.list = monitor.params)
}
}

