% Generated by roxygen2 (4.0.2): do not edit by hand
\name{nameCoefs}
\alias{nameCoefs}
\title{nameCoefs: Match names of coefficients to the parameter names generated by coda}
\usage{
nameCoefs(coef.summary, rand.levels, family, conditional = TRUE,
  form = NULL, name = NULL)
}
\arguments{
\item{coef.summary}{Dataframe generated with the avgCoefs function}

\item{rand.levels}{Levels of the random effects factor (e.g. levels(as.factor(data$site)))}

\item{family}{Character string of the coda group names to match to the original names (e.g. "B.site")}

\item{conditional}{Logical if true match the names to the conditional (specific) estimates for the random effects. If false match the names with the mean summary effects.}

\item{form}{Formula object used for fitting the random effects indicated in family (e.g. formulae$site.form)}

\item{name}{Optional character string to rename the coefficient name column}
}
\description{
\code{nameCoefs} returns a dataframe of the coefficient summary plus the original names
}
\details{
var: blah, blah, blah
value: something, something
}

