% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Summary_Functions.R
\name{nse}
\alias{nse}
\title{nse: Nash-Sutcliffe Efficiency}
\usage{
nse(obs, pred, warn = FALSE)
}
\arguments{
\item{obs}{numeric vector of observed values}

\item{pred}{numeric vector of predicted values with the same length as obs}

\item{warn}{logical whether to print or supress the warning message when NSE not calculated}
}
\description{
\code{nse} returns Nash-Sutcliffe Efficiency coefficient
}
\details{
var: blah, blah, blah
value: something, something
}

