% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plotPredict}
\alias{plotPredict}
\title{Plot observed and predicted values of stream temperature}
\usage{
plotPredict(observed, predicted, siteList = "ALL", yearList = "ALL",
  dir = getwd(), display = FALSE)
}
\arguments{
\item{observed}{Dataframe of observed stream temperatures with at least columns "temp", "sites", and "years"}

\item{predicted}{Dataframe of predicted stream temperatures with at least columns "tempPredicted", "sites", "years", and "airTemp"}

\item{siteList}{Optional character vector of site names for predictions. Default "ALL" will predict to all sites in the predicted dataframe}

\item{yearList}{Optional character vector of years for predictions. Default "ALL" will predict for all years in the predicted dataframe}

\item{dir}{Directory where the files will be saved. Defaults to the working directory}
}
\value{
Saves a png file for every site in the siteList to the dir with all years in the yearList
}
\description{
\code{plotPredicted} plot observed and predicted values of stream temperature along with daily air temperature
}
\details{
blah, blah, blah
}
\examples{
\dontrun{
plotPredict(observed = tempDataSync, predicted = tempFull, siteList = "ALL", yearList = "ALL", dir = paste0(dataLocalDir,'/', 'plots/fullRecord/'))
}
}

