% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plotRawTemperatureData}
\alias{plotRawTemperatureData}
\title{plot raw temps}
\usage{
plotRawTemperatureData(masterData, plotDirectory)
}
\description{
\code{plotRawTemperatureData} returns something of great importance
}
\details{
var: blah, blah, blah
value: something, something

This function creates plots of raw stream temperature data.
It takes the following:
   1) The stream temperature record (unique site ID, date, stream temperature, and air temperature columns)
   2) The directory to save the plots to

It returns a PNG file with plots of air and water plotted against each other and over time.
}

