% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictionFunctions.R
\name{predictTemp}
\alias{predictTemp}
\title{predictTemp: Predict conditional daily temperatures for observed or unobserved sites}
\usage{
predictTemp(data, coef.list, cov.list, rand_ids, Random_AR1 = FALSE)
}
\arguments{
\item{data}{Data frame of covariates for prediction}

\item{coef.list}{List of coefficient values estimated from the model}

\item{cov.list}{List of covariates used in the model}

\item{data.fit}{Data frame used for fitting (calibrating) the JAGS model}

\item{featureid_site}{dataframe with featureid and sites linked}

\item{validate}{Logical if TRUE then the trend is used for the predictions and not adjusted for autocorrelation in the error. Defaults to FALSE}
}
\value{
Numeric vector of predicted daily stream temperatures
}
\description{
\code{predictTemp} Predict daily stream temperatures
}
\details{
The predictions are all conditional on site, HUC8, and year when the data is available and predicts the mean values when any component was not used in the fitting (not in the calibration dataset)
}
\examples{

\dontrun{
Predictions <- predictTemp(data = tempDataSyncValidS, data.fit = tempDataSyncS, cov.list = cov.list, coef.list = coef.list, featureid_site = featureid_site))
}
}

