% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictionFunctions.R
\name{prepConditionalCoef}
\alias{prepConditionalCoef}
\title{prepConditionalCoef}
\usage{
prepConditionalCoef(coef.list, cov.list, var.name)
}
\arguments{
\item{coef.list}{List of coefficient values estimated from the model}

\item{cov.list}{List of covariates used in the model}

\item{var.name}{Name of variable for which coefficients need to be organized}
}
\value{
Returns Dataframe of coefficients from a fitted model associated with a variable (var.name)
}
\description{
\code{prepConditionalCoef} Helper function to prepare dataframe for predictions
}
\details{
Used within the prepPredictDF function
}
\examples{

\dontrun{

}
}

