% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/predictionFunctions.R
\name{prepData}
\alias{prepData}
\title{prepData}
\usage{
prepData(catches_string, springFallBPs, df_covariates_upstream, tempDataSync,
  featureid_lat_lon, featureid_huc8, rand_ids)
}
\arguments{
\item{data}{Dataframe for which predictions will be calculated}

\item{cov.list}{List of covariates used in the model}

\item{coef.list}{List of coefficient values estimated from the model}

\item{featureid_site}{dataframe with featureid and site as 2 columns}
}
\value{
Returns Dataframe of covariates, coefficients from a fitted model, and observed temperature when available
}
\description{
\code{prepData} Helper function to prepare dataframe for predictions
}
\details{
Used within the predictTemp function
}
\examples{
\dontrun{

}
}

