% Generated by roxygen2 (4.0.2): do not edit by hand
\name{prepDataWrapper}
\alias{prepDataWrapper}
\title{prepDataWrapper: Wrapper to prepare data for analysis or predictions}
\usage{
prepDataWrapper(data.fit = NULL, var.names, dataInDir, dataOutDir,
  predict.daymet = FALSE, validate = FALSE, validateFrac = NULL,
  filter.area = NULL)
}
\arguments{
\item{var.names}{Character vector naming the variables to standardize for analysis}
}
\description{
\code{prepDataWrapper} Wrapper to prepare data for analysis or predictions
}
\details{
var: blah, blah, blah
value: something, something
}

