% Generated by roxygen2 (4.0.2): do not edit by hand
\name{prepPredictDF}
\alias{prepPredictDF}
\title{prepPredictDF}
\usage{
prepPredictDF(data, coef.list, cov.list, var.name, featureid_site)
}
\arguments{
\item{data}{Dataframe for which predictions will be calculated}

\item{coef.list}{List of coefficient values estimated from the model}

\item{cov.list}{List of covariates used in the model}
}
\value{
Returns Dataframe of covariates, coefficients from a fitted model, and observed temperature when available
}
\description{
\code{prepPredictDF} Helper function to prepare dataframe for predictions
}
\details{
Used within the predictTemp function
}
\examples{
\dontrun{

}
}

