% Generated by roxygen2 (4.0.1): do not edit by hand
\name{readStreamTempData}
\alias{readStreamTempData}
\title{readStreamTempData}
\usage{
readStreamTempData(timeSeries, covariates, dataSourceList, fieldListTS,
  fieldListCD, directory)
}
\arguments{
\item{timeSeries}{A TRUE/FALSE statement of whether to read in the timeseries data.}

\item{covariates}{A TRUE/FALSE statement of whether to read in the covariate data.}

\item{dataSourceList}{A character vector of the agency abbreviations of the data sources.}

\item{fieldListTS}{A character vector of the common fields to be in the output dataframe.}

\item{fieldListCD}{A character vector of the common fields to be in the output dataframe. If set equal to 'ALL' then all fields are selected.}

\item{directory}{A character vector of the parent dataframe of where the data is stored (dataIn).}
}
\value{
Returns a dataframe with the site name, lat/lon, FEATUREID, and the select covariate values
}
\description{
\code{readStreamTempData} reads in stream temp timeseries and/or the respective covariate data for different data sources, joins them together, and outputs a dataframe
}
\details{
This function reads in stream temp timeseries and/or the respective covariate data for different data sources, joins them together, and outputs a dataframe.
}

