% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rmse}
\alias{rmse}
\title{rmse: root mean squared error}
\usage{
rmse(error)
}
\arguments{
\item{error}{Vector of residual error from a model}
}
\description{
\code{rmse} returns root mean squared error
}
\details{
var: blah, blah, blah
value: something, something
}

