% Generated by roxygen2 (4.0.1): do not edit by hand
\name{sampleJagsInChunks}
\alias{sampleJagsInChunks}
\title{Sample JAGS in Chuncks}
\usage{
sampleJagsInChunks(cl, fileOutName, bugsName, data, inits, nAdapt, params,
  nIterChunk, nThin)
}
\arguments{
\item{cl}{a cluster object, created by parallel package or by package snow. If NULL, use the registered default cluster.}

\item{fileOutName}{Name of output file}

\item{bugsName}{JAGS model name?}

\item{data}{List of data for rjags}

\item{inits}{Function for initializing starting values}

\item{nAdapt}{Length of adaptation/burnin phase}

\item{params}{List of parameters to monitor}

\item{nIterChunk}{Number of iterations in each chunk}

\item{nThin}{Thinning rate (not actually a rate)}
}
\description{
\code{stdCovs} samples jags in parallel chunks with rjags jags.samples instead of coda.samples.
}
\details{
blah, blah, blah
}

