% Generated by roxygen2 (4.0.2): do not edit by hand
\name{stdCovs}
\alias{stdCovs}
\title{Standardize variables using the mean and standard deviation from the analysis}
\usage{
stdCovs(x, y, varNames)
}
\description{
\code{stdCovs} returns the standardized values of all the variables in x.
}
\details{
x: data frame of covariates for prediction or validation
y: original data frame of covariates used in the analysis
}

