% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stdCovs.R
\name{stdFitCovs}
\alias{stdFitCovs}
\title{Standardize covariates for model fitting}
\usage{
stdFitCovs(x, var.names)
}
\arguments{
\item{x}{Dataframe of covariates for prediction or validation}

\item{var.names}{Character vector of covariate names common in x and y to standardize}
}
\description{
\code{stdFitCovs} returns a dataframe of the standardized values of all the variables in x.
}
\details{
This function standardizes the list of covariates in the new dataframe (x) using the means and standard deviations from the dataframe (y) that was used to fit (calibrate) the model
}

