% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{streamTemperature}
\alias{streamTemperature}
\alias{streamTemperature-package}
\title{streamTemperature.}
\description{
streamTemperature.

This package cleans daily and subdaily temperature data, reduces subdaily to daily data, joins with Daymet and landscape data. It also uses a hierarchical model in JAGS to estimate coefficients and predicted daily temperature.

TEXT AS EXAMPLE OF LINKS: The only function you're likely to need from \pkg{roxygen2} is
\code{\link{roxygenize}}. Otherwise refer to the vignettes to see
how to format the documentation.
}

