% Generated by roxygen2 (4.0.2): do not edit by hand
\name{validateModel}
\alias{validateModel}
\title{validateModel: Leave 1 site out model validation}
\usage{
validateModel(model, d, vd)
}
\description{
\code{validateModel} returns something of great importance
}
\details{
var: blah, blah, blah
value: something, something
}

