% Generated by roxygen2 (4.0.2): do not edit by hand
\name{calcYearsMaxTemp}
\alias{calcYearsMaxTemp}
\title{calcYearsMaxTemp}
\usage{
calcYearsMaxTemp(grouped.df, derived.df, temp.threshold)
}
\arguments{
\item{grouped.df}{Dataframe grouped by site then year}

\item{derived.df}{Dataframe of derived metrics}

\item{temp.threshold}{Optional numeric temperature threshold value in degrees C}
}
\value{
Returns Dataframe of derived metrics (e.g. max temperature predicted over daymet record) for each site across all years
}
\description{
\code{calcYearsMaxTemp} Calculate derived metrics from predicted stream temperatures
}
\details{
blah, blah, blah, something, something
}
\examples{
\dontrun{

}
}

