% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Derive_Metrics.R
\name{deriveMetrics}
\alias{deriveMetrics}
\title{deriveMetrics}
\usage{
deriveMetrics(data)
}
\arguments{
\item{data}{Dataframe created with the prepData and predictTemp functions}
}
\value{
Returns Dataframe of derived metrics (e.g. max temperature predicted over daymet record) for each featureid across all years
}
\description{
\code{deriveMetrics} Wrapper to calculate derived metrics in parallel using foreach %dopar%
}
\details{
blah, blah, blah, something, something
}
\examples{
\dontrun{

}
}

