% Generated by roxygen2 (4.0.2): do not edit by hand
\name{flag_incomplete}
\alias{flag_incomplete}
\title{flag_incomplete}
\usage{
flag_incomplete(data)
}
\arguments{
\item{data}{time series data pull from the postgres database (df_values table)}
}
\value{
Flag days with less than 90% of median number of observations for that series
}
\description{
\code{flag_incomplete} Flag incomplete days
}
\details{
blah, blah, blah, something, something
}
\examples{
\dontrun{

}
}

