% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/QAQC.R
\name{flag_warm_influence}
\alias{flag_warm_influence}
\title{flag_mad}
\usage{
flag_warm_influence(data, vals, refs, threshold)
}
\arguments{
\item{data}{time series data pull from the postgres database (df_values table)}

\item{vals}{Character name of the column containing the time series values to test}

\item{refs}{Character name of the column containing the reference time series values}

\item{threshold}{Numeric proportion (between 0-1) indicating the percent of test values above the corresponding reference values resulting in a TRUE flag}

\item{data}{dataframe with two columns of time series}
}
\value{
Median observations per day for each time series

dataframe input data returned with additional logical column indicating whether the time series is excessively warm compared to the reference
}
\description{
\code{dewater} Median observations per day for each time series
}
\details{
blah, blah, blah, something, something
}
\examples{
\dontrun{

}
}

