% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/QAQC.R
\name{flag_warm_influence}
\alias{flag_warm_influence}
\title{flag_warm_influence}
\usage{
flag_warm_influence(data, vals, refs, threshold)
}
\arguments{
\item{data}{dataframe with two columns of time series}

\item{vals}{Character name of the column containing the time series values to test}

\item{refs}{Character name of the column containing the reference time series values}

\item{threshold}{Numeric proportion (between 0-1) indicating the percent of test values above the corresponding reference values resulting in a TRUE flag}
}
\value{
dataframe input data returned with additional logical column indicating whether the time series is excessively warm compared to the reference
}
\description{
\code{flag_warm_influence} Flag series with excessive number of values above reference series
}
\details{
The original intended use of this function is to identify stream sites with significant influence from warm impoundments (spillover dams or beaver dams). If the water temperature is frequently higher than the air temperature on that day and location, there is likely an upstream impoundment affecting the thermal regime at the point of the temperature logger.
}
\examples{
\dontrun{

}
}

