% Generated by roxygen2 (4.0.1): do not edit by hand
\name{makeBiasMap}
\alias{makeBiasMap}
\title{makeBiaseMap}
\usage{
makeBiasMap(d)
}
\description{
\code{makeBiaseMap} returns something of great importance
}
\details{
requires data produced from validateModel()
var: blah, blah, blah
value: something, something
}

