% Generated by roxygen2 (4.0.2): do not edit by hand
\name{obs_freq}
\alias{obs_freq}
\title{obs_freq}
\usage{
obs_freq(data)
}
\arguments{
\item{data}{time series data pull from the postgres database (df_values table)}
}
\value{
Median observations per day for each time series
}
\description{
\code{obs_freq} Median observations per day for each time series
}
\details{
blah, blah, blah, something, something
}
\examples{
\dontrun{

}
}

