% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/QAQC.R
\name{obs_freq}
\alias{obs_freq}
\title{dewater}
\usage{
obs_freq(data)
}
\arguments{
\item{data}{time series data pull from the postgres database (df_values table)}

\item{data}{time series data pull from the postgres database (df_values table)}
}
\value{
Median observations per day for each time series

Median observations per day for each time series
}
\description{
\code{dewater} Median observations per day for each time series
}
\details{
blah, blah, blah, something, something
}
\examples{
\dontrun{

}
}

