% Generated by roxygen2 (4.0.2): do not edit by hand
\name{predictTemp}
\alias{predictTemp}
\title{predictTemp: Predict conditional daily temperatures for observed or unobserved sites}
\usage{
predictTemp(data, data.fit = tempDataSyncS, coefs, firstObsRows, evalRows,
  observed = TRUE)
}
\arguments{
\item{data}{Data frame of covariates for prediction}

\item{data.fit}{Data frame used for fitting (calibrating) the JAGS model}

\item{firstObsRows}{Dataframe with the rowNum column indicating the rows where a logger was first deployed}

\item{evalRows}{Dataframe with the rowNum column indicating the rows during a deployment after the first day for use in the autoregressive}

\item{B.fixed}{Dataframe of fixed effect coefficients}

\item{B.site}{Dataframe of random site coefficients}

\item{B.huc}{Dataframe of random huc coefficients}

\item{B.year}{Dataframe of random year coeffcients}

\item{B.ar1}{Dataframe of AR1 coeffcients by random site}
}
\value{
Numeric vector of predicted daily stream temperatures
}
\description{
\code{predictTemp} Predict daily stream temperatures
}
\details{
The predictions are all conditional on site, HUC8, and year when the data is available and predicts the mean values when any component was not used in the fitting (not in the calibration dataset)
}
\examples{
\dontrun{
Predictions <- predictTemp(data = tempDataSyncValidS, data.fit = tempDataSyncS, firstObsRows = firstObsRows, evalRows = evalRows, B.fixed = B.fixed, B.site = B.site, B.huc = B.huc, B.year = B.year, B.ar1 = B.ar1))
}
}

