% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/predictionFunctions.R
\name{predictTemp}
\alias{predictTemp}
\title{predictTemp: Predict conditional daily temperatures for observed or unobserved sites}
\usage{
predictTemp(data, data.fit = tempDataSyncS, coef.list, cov.list,
  featureid_site)
}
\arguments{
\item{data}{Data frame of covariates for prediction}

\item{data.fit}{Data frame used for fitting (calibrating) the JAGS model}

\item{coef.list}{List of coefficient values estimated from the model}

\item{cov.list}{List of covariates used in the model}

\item{featureid_site}{dataframe with featureid and sites linked}
}
\value{
Numeric vector of predicted daily stream temperatures
}
\description{
\code{predictTemp} Predict daily stream temperatures
}
\details{
The predictions are all conditional on site, HUC8, and year when the data is available and predicts the mean values when any component was not used in the fitting (not in the calibration dataset)
}
\examples{
\dontrun{
Predictions <- predictTemp(data = tempDataSyncValidS, data.fit = tempDataSyncS, cov.list = cov.list, coef.list = coef.list, featureid_site = featureid_site))
}
}

