% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Data_Prep_Functions.R, R/predictionFunctions.R
\name{prepData}
\alias{prepData}
\title{prepData: Wrapper to prepare data for analysis or predictions}
\usage{
prepData(catches_string, springFallBPs, df_covariates_upstream, tempDataSync,
  featureid_lat_lon, featureid_huc8, rand_ids)

prepData(catches_string, springFallBPs, df_covariates_upstream, tempDataSync,
  featureid_lat_lon, featureid_huc8, rand_ids)
}
\arguments{
\item{catches_string}{Vector of catchment featureid to prepare for analysis}

\item{springFallBPs}{Dataframe of spring-fall breakpoints}

\item{df_covariates_upstream}{Dataframe of covariates}

\item{tempDataSync}{Data used for modeling in JAGS}

\item{featureid_lat_lon}{Dataframe linking featureid to lat and lon}

\item{featureid_huc8}{Dataframe linking featureid to hucs}

\item{rand_ids}{Dataframe of random site, huc, and year info}

\item{df_stds}{Dataframe of means and standard deviations used for standardizing continuous covariates}

\item{data}{Dataframe for which predictions will be calculated}

\item{cov.list}{List of covariates used in the model}

\item{coef.list}{List of coefficient values estimated from the model}

\item{featureid_site}{dataframe with featureid and site as 2 columns}
}
\value{
Returns Dataframe of covariates, coefficients from a fitted model, and observed temperature when available
}
\description{
\code{prepData} Wrapper to prepare data for analysis or predictions

\code{prepData} Helper function to prepare dataframe for predictions
}
\details{
This function is a mess but currently works. Lots of redundancy, especially in the data imported.
var: blah, blah, blah
value: something, something

Used within the predictTemp function
}
\examples{
\dontrun{

}
}

