% Generated by roxygen2 (4.0.2): do not edit by hand
\name{prepDF}
\alias{prepDF}
\title{prepDF: Prepares dataframe for use or prediction}
\usage{
prepDF(data, form = formulae)
}
\arguments{
\item{data}{Dataframe of potential covariates to model or predict daily stream temperature}

\item{form}{Named list of formulae each created with the formula function. Names must be data.fixed, data.random.sites, and data.random.years.}
}
\value{
List of named data frames each created using model.matrix and the input formulae
}
\description{
\code{prepDF} Prepares dataframe for use or prediction
}
\details{
blah, blah, blah
}
\examples{
\dontrun{
data.list <- prepDF(data = df, form = formulae)
}
}

