% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/temperatureModelingFunctions.R
\name{riverLabeller}
\alias{riverLabeller}
\title{riverLabeller: Label Rivers}
\usage{
riverLabeller(var, value)
}
\description{
\code{riverLabeller} returns something of great importance
}
\details{
var: blah, blah, blah
value: something, something
}

