% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Summary_Functions.R
\name{rmse}
\alias{rmse}
\title{rmse: root mean squared error}
\usage{
rmse(error, na.rm = T)
}
\arguments{
\item{error}{Vector of residual error from a model}
}
\description{
\code{rmse} returns root mean squared error
}
\details{
var: blah, blah, blah
value: something, something
}

