% Generated by roxygen2 (4.0.1): do not edit by hand
\name{stdCovs}
\alias{stdCovs}
\title{Standardize variables using the mean and standard deviation from the analysis}
\usage{
stdCovs(x, y, var.names)
}
\arguments{
\item{x}{Dataframe of covariates for prediction or validation}

\item{y}{Dataframe used for model fitting}

\item{var.names}{Character vector of covariate names common in x and y to standardize}
}
\description{
\code{stdCovs} returns a dataframe of the standardized values of all the variables in x.
}
\details{
This function standardizes the list of covariates in the new dataframe (x) using the means and standard deviations from the dataframe (y) that was used to fit (calibrate) the model
}

